backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.11"
sqlalchemy = "^2.0.38"
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.1"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.1"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
